test(compaction): consolidate Session integration test into compaction.test.ts
The Session-level integration test for issue #120 was initially placed in a
separate file (test/session-compaction.test.ts), but that duplicated the
@mariozechner/pi-ai mock that compaction.test.ts already owns. Moving the
test into compaction.test.ts under a "Session compaction integration"
describe block reuses the existing pi-ai mock and keeps all compaction-
related regression coverage in one place.

No behavior change; same assertions, same verification against pre-fix code.
